// src/app/shared/model/SideMenuItems.ts
export class SideMenuItems{
    menuType: string = '';
    parentMenuName: string = '';
    menuName: string = '';
    menuCondition: string = '';
    displayName: string = '';
    menuText: string = '';
    menuIconName: string = '';
    isDeleted!: boolean;
    sortOrder!: number;
    isActive!:boolean;
}
    